XRDDEV-746 Move and rename views
